convert other templates
