+++ example-site/components/FormExampleMultiStep/FormExampleMultiStep.tsx
@@ -186,36 +186,31 @@ export const FormExampleMultiStep = () => {
 	};
 
 	return (
-		<AppLayout
-			template={{ name: 'Multi-page form', slug: 'multi-page-form' }}
-			focusMode
-		>
-			<PageContent>
-				<context.Provider value={contextValue}>
-					<Columns>
-						<Column columnSpan={{ xs: 12, md: 4, lg: 3 }}>
-							<ContentBleed visible={{ md: false }}>
-								<ProgressIndicator
-									items={FORM_STEPS.map(({ label }, idx) => ({
-										label,
-										status: getStepStatus(idx),
-										onClick: () => setCurrentStep(idx),
-									}))}
-								/>
-							</ContentBleed>
-						</Column>
-						<Column columnSpan={{ xs: 12, md: 8 }} columnStart={{ lg: 5 }}>
-							<Stack gap={3} alignItems="flex-start">
-								<DirectionButton direction="left" onClick={back}>
-									Back
-								</DirectionButton>
-								{FormStepComponent ? <FormStepComponent /> : null}
-							</Stack>
-						</Column>
-					</Columns>
-				</context.Provider>
-			</PageContent>
-		</AppLayout>
+		<PageContent>
+			<context.Provider value={contextValue}>
+				<Columns>
+					<Column columnSpan={{ xs: 12, md: 4, lg: 3 }}>
+						<ContentBleed visible={{ md: false }}>
+							<ProgressIndicator
+								items={FORM_STEPS.map(({ label }, idx) => ({
+									label,
+									status: getStepStatus(idx),
+									onClick: () => setCurrentStep(idx),
+								}))}
+							/>
+						</ContentBleed>
+					</Column>
+					<Column columnSpan={{ xs: 12, md: 8 }} columnStart={{ lg: 5 }}>
+						<Stack gap={3} alignItems="flex-start">
+							<DirectionButton direction="left" onClick={back}>
+								Back
+							</DirectionButton>
+							{FormStepComponent ? <FormStepComponent /> : null}
+						</Stack>
+					</Column>
+				</Columns>
+			</context.Provider>
+		</PageContent>
 	);
 };
