+++ example-site/components/templates/Category.stories.tsx
@@ -1,0 +1,14 @@
+import { PageTemplate } from '../../../.storybook/components/PageTemplate';
+import { Category as CategoryContent } from './Category';
+
+export default {
+	title: 'Templates/Category',
+};
+
+export function Category() {
+	return (
+		<PageTemplate>
+			<CategoryContent />
+		</PageTemplate>
+	);
+}
