+++ example-site/components/templates/Content.stories.tsx
@@ -1,0 +1,14 @@
+import { PageTemplate } from '../../../.storybook/components/PageTemplate';
+import { Content as ContentComponent } from './Content';
+
+export default {
+	title: 'Templates/Content',
+};
+
+export function Content() {
+	return (
+		<PageTemplate>
+			<ContentComponent />
+		</PageTemplate>
+	);
+}
